Split Engrave into base implementation and Engrave specific pieces.
